feat(brand): add capybara logo to README and dashboard UI
- docs/assets/cap-evolve-logo.png (full-res) shown at the top of the README
- dashboard/frontend/public: 256px logo + 64px favicon (downscaled from 1254px)
- BrandLogo component renders the logo with the live-pulse ring; replaces the
  hand-drawn Capybara SVG in AppShell (sidebar + mobile top bar)
- favicon switched to the logo
Frontend build green; 32 vitest pass; live server serves /cap-evolve-logo.png (200).
